requirements.txt
# Core ML Libraries
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
tensorflow>=2.15.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Model Persistence
joblib==1.3.2

# Web Framework
streamlit==1.29.0

# User Authentication
streamlit-authenticator==0.2.3
bcrypt==4.1.2
PyYAML==6.0.1

# Database
sqlalchemy==2.0.25

# Explainable AI
shap==0.44.1
lime==0.2.0.1

# Utilities
python-dateutil==2.8.2
